+++ R/rbind_labelled.R
@@ -71,16 +71,7 @@ rbind_labelled <- function(..., labels=NULL, warn=TRUE) {
   }
 
   # what kind of dataset is it we are working with
-  if (is.element("label.table", attributes(dfs[[1]]) %>% names())) {
-    type <- "foreign"
-  } else if (any(lapply(dfs[[1]], class) %>% unlist() == "labelled")) {
-    type <- "labelled"
-  } else if (any(lapply(dfs[[1]], attributes) %>%
-                 lapply(names) %>% unlist() == "label")) {
-    type <- "reformat"
-  } else {
-    type <- "reformat"
-  }
+  type <- dataset_label_type(dfs[[1]], stop = FALSE)
 
   # if the data has names let's grab that and append it
   if (!is.null(names(dfs))) {
@@ -91,15 +82,15 @@ rbind_labelled <- function(..., labels=NULL, warn=TRUE) {
   }
 
   # if its a foregin or reformatted dataset list then skip over the labelling
-  if (!is.element(type, c("foreign", "reformat"))) {
+  if (type == "labelled") {
 
     ## Ensure same column ordering for all dfs
     dfs <- lapply(dfs, "[", names(dfs[[1]]))
 
     # check to see if variables are labelled
     islab <- vapply(dfs, function(x){
-        vapply(x, haven::is.labelled, logical(1))
-      }, logical(length(dfs[[1]])))
+      vapply(x, haven::is.labelled, logical(1))
+    }, logical(length(dfs[[1]])))
 
     ## let's catch for one variable dataframes
     islab_vec_catch <- function(islab_obj) {
@@ -144,7 +135,7 @@ rbind_labelled <- function(..., labels=NULL, warn=TRUE) {
       }
 
       lablist <- lapply(dfs, "[", intersect(anylab, needslab))
-      lablist <- lapply(lablist, lapply, attr, "labels")
+      lablist <- lapply(lablist, lapply, attr, "labels", exact = TRUE)
       lablist <- do.call(Map, c(f = list, lablist))
 
       allequal <- lapply(lablist, function(x) {
@@ -170,7 +161,8 @@ rbind_labelled <- function(..., labels=NULL, warn=TRUE) {
                         function(x) min(which(x)))
       if (length(whichlab)) {
         inherlab <- setNames(
-          lapply(Map("[[", dfs[whichlab], names(whichlab)), attr, "labels"),
+          lapply(Map("[[", dfs[whichlab], names(whichlab)),
+                 attr, "labels", exact = TRUE),
           names(whichlab)
         )
         labels <- c(labels, inherlab)
@@ -180,13 +172,29 @@ rbind_labelled <- function(..., labels=NULL, warn=TRUE) {
     ## rbind data frames
     df <- do.call(rbind, dfs)
 
-    ## Add labels
-    df[names(labels)] <- Map(haven::labelled, df[names(labels)], labels)
-
     ## Convert concatenated variables back to labelled
     df[catvar] <- lapply(df[catvar], factor)
-    df[catvar] <- lapply(df[catvar], function(x)
-      haven::labelled(as.integer(x), setNames(seq_along(levels(x)), levels(x))))
+
+    # match on haven package version
+    if (packageVersion("haven") > "1.1.2") {
+
+      ## Get the label attribute
+      labl <- lapply(df[names(labels)], attr, "label", exact = TRUE)
+
+      ## Create labelled class with the label attributes sved
+      df[names(labels)] <- Map(haven::labelled, df[names(labels)], labels, labl)
+      df[catvar] <- lapply(df[catvar], function(x) {
+        haven::labelled(as.integer(x),
+                        setNames(seq_along(levels(x)), levels(x)),
+                        attr(x, "label", exact = TRUE))})
+    } else {
+
+      ## Create labelled class
+      df[names(labels)] <- Map(haven::labelled, df[names(labels)], labels)
+      df[catvar] <- lapply(df[catvar], function(x) {
+        haven::labelled(as.integer(x),
+                        setNames(seq_along(levels(x)), levels(x)))})
+    }
 
   } else {
 
@@ -196,4 +204,80 @@ rbind_labelled <- function(..., labels=NULL, warn=TRUE) {
   }
 
   return(df)
+}
+
+
+
+#' Archived dataset capable as_factor
+#'
+#' @description
+#' Changes in `haven` have meant that `labelled` class are now referred to as
+#' `haven_labelled` classes. If `haven::as_factor` is used on old datasets they
+#' will fail to find the suitable method. rdhs::as_factor.labelled will work on
+#' old archived datasets that have a `labelled` class.
+#'
+#' @details
+#' For more details see \code{haven::as_factor}
+#'
+#' @inheritParams haven::as_factor
+#' @export
+#'
+#' @examples
+#'
+#' # create a data.frame using the new haven_labelled class
+#' df1 <- data.frame(
+#' area = haven::labelled(c(1L, 2L, 3L), c("reg 1"=1,"reg 2"=2,"reg 3"=3)),
+#' climate = haven::labelled(c(0L, 1L, 1L), c("cold"=0,"hot"=1))
+#' )
+#'
+#' # manually change it to the old style
+#' class(df1$area) <- "labelled"
+#' class(df1$climate) <- "labelled"
+#'
+#' # with rdhs attached, i.e. library(rdhs), we can now do the following
+#' haven::as_factor(df1$area)
+#'
+#' # we can also use this on the data.frame by using the only_labelled argument
+#' haven::as_factor(df1, only_labelled = TRUE)
+
+as_factor.labelled <- function(x,
+                               levels = c("default", "labels", "values", "both"),
+                               ordered = FALSE, ...) {
+
+  replace_with <- function(x, from, to) {
+    stopifnot(length(from) == length(to))
+    out <- x
+    matches <- match(x, from, incomparables = NA)
+    out[!is.na(matches)] <- to[matches[!is.na(matches)]]
+    tagged <- haven::is_tagged_na(x)
+    if (!any(tagged)) {
+      return(out)
+    }
+    matches <- match(haven::na_tag(x), haven::na_tag(from), incomparables = NA)
+    out[!is.na(matches)] <- to[matches[!is.na(matches)]]
+    out
+  }
+
+  levels <- match.arg(levels)
+  label <- attr(x, "label", exact = TRUE)
+  labels <- attr(x, "labels")
+  if (levels == "default" || levels == "both") {
+    if (levels == "both") {
+      names(labels) <- paste0("[", labels, "] ", names(labels))
+    }
+    vals <- unique(x)
+    levs <- replace_with(vals, unname(labels), names(labels))
+    levs <- sort(c(stats::setNames(vals, levs), labels),
+                 na.last = TRUE)
+    levs <- unique(names(levs))
+    x <- replace_with(x, unname(labels), names(labels))
+    x <- factor(x, levels = levs, ordered = ordered)
+  }
+  else {
+    levs <- unname(labels)
+    labs <- switch(levels, labels = names(labels), values = levs)
+    x <- replace_with(x, levs, labs)
+    x <- factor(x, labs, ordered = ordered)
+  }
+  structure(x, label = label)
 }